# default.openapi .json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Default module",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/purchases": {
      "post": {
        "summary": "create",
        "deprecated": false,
        "description": "",
        "tags": [
          "purchases"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              },
              "example": {
                "supplierId": 1
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "message": "Purchase created"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "get": {
        "summary": "get-all",
        "deprecated": false,
        "description": "",
        "tags": [
          "purchases"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "",
            "required": false,
            "example": 1,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": false,
            "example": 1,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "",
            "required": false,
            "example": "name",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              },
              "examples": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "message": "Purchases fetched successfully",
                  "data": [
                    {
                      "purchase_id": 10,
                      "supplie_id": 1,
                      "supplier_name": "ABC Farm",
                      "purchase_date": "2026-03-07",
                      "total_cost": 1675
                    },
                    {
                      "purchase_id": 10,
                      "supplier_id": 1,
                      "supplier_name": "ABC Farm",
                      "purchase_date": "2026-03-07",
                      "total_cost": 1675
                    },
                    {
                      "purchase_id": 10,
                      "supplier_id": 1,
                      "supplier_name": "ABC Farm",
                      "purchase_date": "2026-03-07",
                      "total_cost": 1675
                    }
                  ],
                  "pagination": {
                    "page": 1,
                    "limit": 10,
                    "total": 3,
                    "total_pages": 1
                  }
                }
              }
            },
            "headers": {}
          },
          "400": {
            "$ref": "#/components/responses/400 Bad Request",
            "description": ""
          },
          "401": {
            "$ref": "#/components/responses/401 Unauthorized",
            "description": ""
          },
          "500": {
            "$ref": "#/components/responses/500 Internal Server Error",
            "description": ""
          }
        },
        "security": []
      }
    },
    "/purchases/{id}": {
      "get": {
        "summary": "get-one",
        "deprecated": false,
        "description": "",
        "tags": [
          "purchases"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "message": "Purchase fetched successfully",
                  "data": {
                    "purchase_id": 10,
                    "supplier_id": 1,
                    "supplier_name": "ABC Farm",
                    "purchase_date": "2026-03-07",
                    "items": [
                      {
                        "purchase_item_id": 1,
                        "stock_id": 1,
                        "stock_name": "Soybean",
                        "quantity": 50,
                        "price": 20,
                        "total": 1000
                      },
                      {
                        "purchase_item_id": 2,
                        "stock_id": 2,
                        "stock_name": "Sugar",
                        "quantity": 30,
                        "price": 10,
                        "total": 300
                      }
                    ],
                    "totalCost": 1675
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/purchases/{id}/items": {
      "get": {
        "summary": "get-item",
        "deprecated": false,
        "description": "",
        "tags": [
          "purchases"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "purchase_id": 10,
                  "items": [
                    {
                      "purchase_item_id": 1,
                      "stock_id": 1,
                      "quantity": 50,
                      "price": 20
                    }
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/purchases/{purchase_id}/items": {
      "post": {
        "summary": "create",
        "deprecated": false,
        "description": "",
        "tags": [
          "purchases"
        ],
        "parameters": [
          {
            "name": "purchase_id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              },
              "example": {
                "items": [
                  {
                    "stock_id": 1,
                    "quantity": 50,
                    "price": 20
                  },
                  {
                    "stock_id": 2,
                    "quantity": 30,
                    "price": 10
                  },
                  {
                    "stock_id": 3,
                    "quantity": 15,
                    "price": 25
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "examples": {
                  "1": {
                    "summary": "success",
                    "value": {
                      "message": "Items added successfully",
                      "data": {
                        "purchase_id": 10,
                        "items_added": [
                          {
                            "purchase_item_id": 1,
                            "stock_id": 1,
                            "quantity": 50,
                            "price": 20,
                            "total": 1000
                          },
                          {
                            "purchase_item_id": 2,
                            "stock_id": 2,
                            "quantity": 30,
                            "price": 10,
                            "total": 300
                          },
                          {
                            "purchase_item_id": 3,
                            "stockId": 3,
                            "quantity": 15,
                            "price": 25,
                            "total": 375
                          }
                        ],
                        "totalCost": 1675
                      }
                    }
                  },
                  "2": {
                    "summary": "some success",
                    "value": {
                      "message": "ข้าวสวย,ถั่วเหลือง มีอยู่เเล้ว",
                      "data": {
                        "purchaseId": 10,
                        "itemsAdded": [
                          {
                            "purchaseItemId": 3,
                            "stockId": 3,
                            "quantity": 15,
                            "price": 25,
                            "total": 375
                          }
                        ],
                        "totalCost": 1675
                      }
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}